TA-SSL-RF/requirements.txt
# 先安装internimage需要的环境，在安装此requirement.txt,然后再改timm库
#mmcv-full==1.6.1
#torch==1.11.0+cu113
#torchvision==0.12.0+cu113

timm==0.6.11
pandas
mmsegmentation==0.27.0
matplotlib
scikit-learn
seaborn
pillow
opencv-python
tqdm
argparse
seaborn
tifffile
typing
termcolor
datetime
wandb
einops
jupyter
tensorboard
thop
torchmetrics
mmengine
lightning-bolts==0.5.0
pytorch_lightning==1.6.4
albumentations==1.2.1
ml-collections==0.1.1
h5py==3.7.0

